# api/openapi.yaml
openapi: 3.1.0
info:
  title: Sheily local sovereign agent
  description: Agente cognitivo local con autoridades independientes y ejecución Rust.
  version: 0.3.0
paths:
...
  /v1/auth/login:
    post:
      summary: Login
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me:
    get:
      summary: Me
      operationId: me_v1_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/conversations:
    post:
      summary: Create Conversation
      operationId: create_conversation_v1_conversations_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/conversations/{conversation_id}/messages:
    get:
      summary: List Messages
      operationId: list_messages_v1_conversations__conversation_id__messages_get
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
                title: Response List Messages V1 Conversations  Conversation Id  Messages
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Message
      operationId: create_message_v1_conversations__conversation_id__messages_post
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/missions/{mission_id}:
    get:
      summary: Get Mission
      operationId: get_mission_v1_missions__mission_id__get
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/missions/{mission_id}/approval:
    post:
      summary: Approve Mission
      operationId: approve_mission_v1_missions__mission_id__approval_post
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mission'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/missions/{mission_id}/events:
    get:
      summary: Stream Events
      operationId: stream_events_v1_missions__mission_id__events_get
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      - name: after
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: After
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/documents:
    post:
      summary: Upload Document
      operationId: upload_document_v1_documents_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_v1_documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPublic'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/memories:
    get:
      summary: List Memories
      operationId: list_memories_v1_memories_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemoryRecord'
                title: Response List Memories V1 Memories Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/memories/{memory_id}:
    delete:
      summary: Delete Memory
      operationId: delete_memory_v1_memories__memory_id__delete
      parameters:
      - name: memory_id
        in: path
        required: true
        schema:
          type: string
          title: Memory Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_document_v1_documents_post:
      properties:
        upload:
          type: string
          contentMediaType: application/octet-stream
          title: Upload
      type: object
      required:
      - upload
      title: Body_upload_document_v1_documents_post
...
    ConversationCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
          default: Nueva conversación
      additionalProperties: false
      type: object
      title: ConversationCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 128
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 512
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    Principal:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          enum:
          - user
          - operator
          - admin
          title: Role
      additionalProperties: false
      type: object
      required:
      - user_id
      - username
      - role
      title: Principal
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user_id:
          type: string
          title: User Id
        role:
          type: string
          title: Role
      additionalProperties: false
      type: object
      required:
      - access_token
      - expires_in
      - user_id
      - role
      title: TokenResponse
...